requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
asyncio==3.4.3
asynctest==0.13.0
attrs==23.1.0
beautifulsoup4==4.9.3
certifi==2023.5.7
chardet==4.0.0
charset-normalizer==3.1.0
fake-useragent==1.1.3
frozenlist==1.3.3
idna==2.10
importlib-metadata==6.7.0
multidict==6.0.4
numpy==1.21.6
pandas==1.2.4
python-dateutil==2.8.2
pytz==2023.3
requests==2.25.1
six==1.16.0
soupsieve==2.4.1
typing_extensions==4.6.3
urllib3==1.26.16
yarl==1.9.2
zipp==3.15.0
asyncio==3.4.3
fake-useragent==1.1.3
apache-airflow-providers-amazon

